Emails a formatted excel spreadsheet
Problem: Users are spending too much time creating pivot tables from our
ERP system for a daily report. However, in order to automate this task
we must take into consideration business days (including holidays) and a
variable budget that needs to be entered by the user.

Solution: A C# console application that is able to do all of the above
using EPPlus, a custom DLL for holidays with an XML file source and a
settings file that can be modified by an end user with little training.